requirements.txt
Flask==0.12.1
Flask-RESTful==0.3.5
pytest==3.0.7
selenium
requests
apscheduler
docutils
flask_apscheduler
IPy
sqlalchemy
psycopg2-binary
python-dotenv